deploy/requirements.txt
pandas
streamlit==1.38.0
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.8.4